refactor: 생성자 주입 방식으로 변경, 같은 service를 참조하지 않도록 변경

+++ src/main/java/nextstep/subway/application/LineService.java
@@ -8,7 +8,6 @@ import nextstep.subway.dto.LineRequest;
 import nextstep.subway.dto.LineResponse;
 import nextstep.subway.dto.LineUpdateRequest;
 import nextstep.subway.exception.NoSuchElementFoundException;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -21,17 +20,17 @@ import java.util.stream.Collectors;
 public class LineService {
 
     private final LineRepository lineRepository;
-    @Autowired
-    private StationService stationService;
+    private final StationRepository stationRepository;
 
-    public LineService(LineRepository lineRepository) {
+    public LineService(LineRepository lineRepository, StationRepository stationRepository) {
         this.lineRepository = lineRepository;
+        this.stationRepository = stationRepository;
     }
 
     @Transactional
     public LineResponse saveLine(LineRequest lineRequest) {
-        Station upStation = stationService.findById(lineRequest.getUpStationId());
-        Station downStation = stationService.findById(lineRequest.getDownStationId());
+        Station upStation = findStationById(lineRequest.getUpStationId());
+        Station downStation = findStationById(lineRequest.getDownStationId());
 
         Line persistLine = lineRepository.save(new Line(lineRequest.getName(), lineRequest.getColor(), upStation, downStation));
         return LineResponse.of(persistLine);
@@ -58,6 +57,13 @@ public class LineService {
 
     @Transactional
     public void deleteLineById(@PathVariable Long id) {
+        if (!lineRepository.existsById(id)) {
+            throw new NoSuchElementFoundException("해당 노선을 찾을 수 없습니다.");
+        }
         lineRepository.deleteById(id);
+    }
+
+    private Station findStationById(Long id) {
+        return stationRepository.findById(id).orElseThrow(() -> new NoSuchElementFoundException("해당 역을 찾을 수 없습니다."));
     }
 }
